Use namespace environment instead of test mode
Test mode is not supported and the watchdogd maintainer has expressed
intention to remove it. The mount namespace with tmpfs allows the test
watchdogd instance to create its files in /var/run without real root
access. Conveniently watchdogd does not fail when the watchdog device
is a regular file, note that if running as root it would directly
trigger reboot if the watchdog device does not.